feat: Refactor AdminLayout to use AdminLayoutClient and simplify structure
