+++ src/app/layout.tsx
@@ -30,7 +30,7 @@ export default function RootLayout({
 
   return (
     <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`} suppressHydrationWarning={true}>
         <Toaster />
         <Suspense fallback={<LoadingPage />}>
           {children}
